requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
Babel==2.12.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-apscheduler==0.6.2
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
phonenumbers==8.13.13
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
